Changing around where the hash object is stored
Hash object has been moved to the core as a node (note, users must add the hash entry in the config under the nodes list for the updated code to work.
Also doing some more typecasting in the authorize node for the heck of it.  Probably don't even need it all anyways, but meh.